samsung-framework: update sources to `S936USQU5ZYH6`
* from one ui 8 beta 5

+++ src/android/hardware/CameraSessionStats.java
@@ -130,8 +130,8 @@ public class CameraSessionStats implements Parcelable {
         parcel.writeBoolean(this.mUsedZoomOverride);
         parcel.writeInt(this.mSessionIndex);
         this.mCameraExtensionSessionStats.writeToParcel(parcel, 0);
-        parcel.writeInt(this.mMostRequestedFpsRange.getLower().intValue());
-        parcel.writeInt(this.mMostRequestedFpsRange.getUpper().intValue());
+        parcel.writeInt(((Integer) this.mMostRequestedFpsRange.getLower()).intValue());
+        parcel.writeInt(((Integer) this.mMostRequestedFpsRange.getUpper()).intValue());
     }
 
     public void readFromParcel(Parcel parcel) {
